+++ Wrappers/Python/cil/framework/framework.py
@@ -9,7 +9,7 @@
 
 #   http://www.apache.org/licenses/LICENSE-2.0
 
-#   Unless required by applicable law or agreed to in writing, software
+#   Unless required by applicable law or agreed to in writing, software 
 #   distributed under the License is distributed on an "AS IS" BASIS,
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
@@ -132,7 +132,7 @@ class ImageGeometry(object):
                         .format(labels_default,labels))
                     
             self.__dimension_labels = tuple(labels)
-
+        
     def __eq__(self, other):
 
         if not isinstance(other, self.__class__):
@@ -154,6 +154,14 @@ class ImageGeometry(object):
             return True
         
         return False
+
+    @property
+    def dtype(self):
+        return self.__dtype
+
+    @dtype.setter
+    def dtype(self, val):
+        self.__dtype = val           
 
     def __init__(self, 
                  voxel_num_x=0, 
@@ -181,6 +189,7 @@ class ImageGeometry(object):
         self.channel_labels = None
         self.channel_spacing = 1.0
         self.dimension_labels = kwargs.get('dimension_labels', None)
+        self.dtype = kwargs.get('dtype', numpy.float32)
 
     def subset(self, dimensions=None, **kw):
         '''Returns a new sliced and/or reshaped ImageGeometry'''
@@ -280,13 +289,15 @@ class ImageGeometry(object):
 
         return repres
     def allocate(self, value=0, **kwargs):
-        '''allocates an ImageData according to the size expressed in the instance'''
-        if value == 'random_int':
-            dtype = kwargs.get('dtype', numpy.int32)
-            if dtype not in [numpy.int, numpy.int8, numpy.int16, numpy.int32, numpy.int64]:
-                raise ValueError('Expecting int type, got {}'.format(dtype))
-        else:
-            dtype = kwargs.get('dtype', numpy.float32)
+        '''allocates an ImageData according to the size expressed in the instance
+        
+        :param value: accepts numbers to allocate an uniform array, or a string as 'random' or 'random_int' to create a random array or None.
+        :type value: number or string, default None allocates empty memory block, default 0
+        :param dtype: numerical type to allocate
+        :type dtype: numpy type, default numpy.float32
+        '''
+
+        dtype = kwargs.get('dtype', self.dtype)
 
         if kwargs.get('dimension_labels', None) is not None:
             raise ValueError("Deprecated: 'dimension_labels' cannot be set with 'allocate()'. Use 'geometry.set_labels()' to modify the geometry before using allocate.")
@@ -313,25 +324,15 @@ class ImageGeometry(object):
                 if seed is not None:
                     numpy.random.seed(seed)
                 max_value = kwargs.get('max_value', 100)
-                out.fill(numpy.random.randint(max_value,size=self.shape, dtype=dtype))
+                r = numpy.random.randint(max_value,size=self.shape, dtype=numpy.int32)
+                out.fill(numpy.asarray(r, dtype=self.dtype))
             elif value is None:
                 pass
             else:
                 raise ValueError('Value {} unknown'.format(value))
 
         return out
-    # The following methods return 2 members of the class, therefore I 
-    # don't think we need to implement them. 
-    # Additionally using __len__ is confusing as one would think this is 
-    # an iterable. 
-    #def __len__(self):
-    #    '''returns the length of the geometry'''
-    #    return self.length
-    #def shape(self):
-    #    '''Returns the shape of the array of the ImageData it describes'''
-    #    return self.shape
-        
-
+    
 class ComponentDescription(object):
     r'''This class enables the creation of vectors and unit vectors used to describe the components of a tomography system
      '''
@@ -1442,6 +1443,14 @@ class AcquisitionGeometry(object):
             self.__dimension_labels = tuple(val)
 
 
+    @property
+    def dtype(self):
+        return self.__dtype
+
+    @dtype.setter
+    def dtype(self, val):
+        self.__dtype = val       
+
     def __init__(self,
                 geom_type, 
                 dimension=None,
@@ -1457,6 +1466,9 @@ class AcquisitionGeometry(object):
 
         """Constructor method
         """
+
+        # default dtype for the acquisition geometry
+        self.dtype = kwargs.get('dtype', numpy.float32)
 
         #backward compatibility
         new_setup = kwargs.get('new_setup', False)
@@ -1747,17 +1759,10 @@ class AcquisitionGeometry(object):
         
         :param value: accepts numbers to allocate an uniform array, or a string as 'random' or 'random_int' to create a random array or None.
         :type value: number or string, default None allocates empty memory block
-        :param dimension_labels: labels for the dimension axis
-        :type list: default None
         :param dtype: numerical type to allocate
         :type dtype: numpy type, default numpy.float32
         '''
-        if value == 'random_int':
-            dtype = kwargs.get('dtype', numpy.int32)
-            if dtype not in [numpy.int, numpy.int8, numpy.int16, numpy.int32, numpy.int64]:
-                raise ValueError('Expecting int type, got {}'.format(dtype))
-        else:
-            dtype = kwargs.get('dtype', numpy.float32)
+        dtype = kwargs.get('dtype', self.dtype)
 
         if kwargs.get('dimension_labels', None) is not None:
             raise ValueError("Deprecated: 'dimension_labels' cannot be set with 'allocate()'. Use 'geometry.set_labels()' to modify the geometry before using allocate.")
@@ -1784,7 +1789,8 @@ class AcquisitionGeometry(object):
                 if seed is not None:
                     numpy.random.seed(seed)
                 max_value = kwargs.get('max_value', 100)
-                out.fill(numpy.random.randint(max_value,size=self.shape, dtype=dtype))
+                r = numpy.random.randint(max_value,size=self.shape, dtype=numpy.int32)
+                out.fill(numpy.asarray(r, dtype=self.dtype))
             elif value is None:
                 pass
             else:
@@ -1842,7 +1848,9 @@ class DataContainer(object):
 
     @property
     def dtype(self):
-        '''Returns the type of the data array'''
+        '''Returns the dtype of the data array. 
+           If geometry exists, the dtype of the geometry = dtype of the array'''                          
+        self.geometry.dtype = self.array.dtype       
         return self.array.dtype
 
     @property
@@ -1868,9 +1876,13 @@ class DataContainer(object):
         if type(self) is DataContainer:
             self.dimension_labels = dimension_labels
 
-        # finally copy the geometry
+        # finally copy the geometry, and force dtype of the geometry of the data = the dype of the data
         if 'geometry' in kwargs.keys():
             self.geometry = kwargs['geometry']
+            try:
+                self.geometry.dtype = self.dtype            
+            except:
+                pass    
         
     def get_dimension_size(self, dimension_label):
 
@@ -2517,10 +2529,13 @@ class ImageData(DataContainer):
         if not kwargs.get('suppress_warning', False):
             warnings.warn('Direct invocation is deprecated and will be removed in following version. Use allocate from ImageGeometry instead',
               DeprecationWarning)
+
         dtype = kwargs.get('dtype', numpy.float32)
+    
 
         if geometry is None:
-            raise AttributeError("ImageGeometry requires a geometry")
+            raise AttributeError("ImageData requires a geometry")
+            
 
         labels = kwargs.get('dimension_labels', None)
         if labels is not None and labels != geometry.dimension_labels:
@@ -2541,7 +2556,7 @@ class ImageData(DataContainer):
         if array.ndim not in [2,3,4]:
             raise ValueError('Number of dimensions are not 2 or 3 or 4 : {0}'.format(array.ndim))
     
-        super(ImageData, self).__init__(array, deep_copy, geometry=geometry,**kwargs)
+        super(ImageData, self).__init__(array, deep_copy, geometry=geometry, **kwargs)
                                
     def subset(self, dimensions=None, **kw):
         '''returns a subset of ImageData and regenerates the geometry'''
@@ -2564,7 +2579,7 @@ class ImageData(DataContainer):
         try:
             geometry_new = self.geometry.get_slice(channel=channel, vertical=vertical, horizontal_x=horizontal_x, horizontal_y=horizontal_y)
         except ValueError:
-            if kw_force:
+            if force:
                 geometry_new = None
             else:
                 raise ValueError ("Unable to return slice of requested ImageData. Use 'force=True' to return DataContainer instead.")
@@ -2605,6 +2620,7 @@ class AcquisitionData(DataContainer):
         if not kwargs.get('suppress_warning', False):
             warnings.warn('Direct invocation is deprecated and will be removed in following version. Use allocate from AcquisitionGeometry instead',
               DeprecationWarning)
+
         dtype = kwargs.get('dtype', numpy.float32)
 
         if geometry is None:
@@ -2655,7 +2671,6 @@ class AcquisitionData(DataContainer):
             if force:
                 geometry_new = None
             else:
-                print(ve)
                 raise ValueError ("Unable to return slice of requested AcquisitionData. Use 'force=True' to return DataContainer instead.")
 
         #get new data
@@ -2946,6 +2961,7 @@ class VectorData(DataContainer):
 
     def __init__(self, array=None, **kwargs):
         self.geometry = kwargs.get('geometry', None)
+
         dtype = kwargs.get('dtype', numpy.float32)
         
         if self.geometry is None:
@@ -2976,13 +2992,21 @@ class VectorGeometry(object):
     '''Geometry describing VectorData to contain 1D array'''
     RANDOM = 'random'
     RANDOM_INT = 'random_int'
+
+    @property
+    def dtype(self):
+        return self.__dtype
+
+    @dtype.setter
+    def dtype(self, val):
+        self.__dtype = val      
         
     def __init__(self, 
                  length, **kwargs):
         
         self.length = length
         self.shape = (length, )
-
+        self.dtype = kwargs.get('dtype', numpy.float32)
         self.dimension_labels = kwargs.get('dimension_labels', None)
         
     def clone(self):
@@ -3005,9 +3029,20 @@ class VectorGeometry(object):
         return False
 
     def allocate(self, value=0, **kwargs):
-        '''allocates an VectorData according to the size expressed in the instance'''
-        self.dtype = kwargs.get('dtype', numpy.float32)
-        out = VectorData(geometry=self.copy(), dtype=self.dtype)
+        '''allocates an VectorData according to the size expressed in the instance
+        
+        :param value: accepts numbers to allocate an uniform array, or a string as 'random' or 'random_int' to create a random array or None.
+        :type value: number or string, default None allocates empty memory block
+        :param dtype: numerical type to allocate
+        :type dtype: numpy type, default numpy.float32
+        :param seed: seed for the random number generator
+        :type seed: int, default None
+        :param max_value: max value of the random int array
+        :type max_value: int, default 100'''
+
+        dtype = kwargs.get('dtype', self.dtype)
+        # self.dtype = kwargs.get('dtype', numpy.float32)
+        out = VectorData(geometry=self.copy(), dtype=dtype)
         if isinstance(value, Number):
             if value != 0:
                 out += value
@@ -3022,7 +3057,8 @@ class VectorGeometry(object):
                 if seed is not None:
                     numpy.random.seed(seed)
                 max_value = kwargs.get('max_value', 100)
-                out.fill(numpy.random.randint(max_value,size=self.shape))
+                r = numpy.random.randint(max_value,size=self.shape, dtype=numpy.int32)
+                out.fill(numpy.asarray(r, dtype=self.dtype))             
             elif value is None:
                 pass
             else:
@@ -3069,3 +3105,5 @@ class DataOrder():
         else:
             raise ValueError("Expected dimension_label order {0}, got {1}.\nTry using `data.reorder('{2}')` to permute for {2}"
                  .format(order_requested, list(geometry.dimension_labels), engine))
+
+